Tweaked showmemstat. Fixed updater minor error. Fixed some compiler warnings.

+++ MTA10/mods/shared_logic/lua/CLuaFunctionDefs.Event.cpp
@@ -71,7 +71,7 @@ int CLuaFunctionDefs::AddEventHandler ( lua_State* luaVM )
         uint iPos = strPriority.find_first_of ( "-+" );
         if ( iPos != SString::npos )
         {
-            fPriorityMod = atof ( strPriority.SubStr ( iPos ) );
+            fPriorityMod = (float)atof ( strPriority.SubStr ( iPos ) );
             strPriority = strPriority.Left ( iPos );
         }
